pyproject.toml
[project]
name = "spanglish-backend"
version = "0.1.0"
description = "A DRF backend for Spanglish, a language learning app."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.10",
    "ruff>=0.12.8",
    "sqlite-utils>=3.38",
]
